Removed interface LogonDurationDAO
