+++ webcurator-core/src/main/java/org/webcurator/auth/WCTAuthenticationSuccessHandler.java
@@ -21,9 +21,6 @@ import org.springframework.security.authentication.UsernamePasswordAuthenticatio
 import org.springframework.security.core.Authentication;
 import org.springframework.security.web.authentication.AuthenticationSuccessHandler;
 import org.springframework.security.web.authentication.SavedRequestAwareAuthenticationSuccessHandler;
-import org.springframework.security.web.savedrequest.HttpSessionRequestCache;
-import org.springframework.security.web.savedrequest.RequestCache;
-import org.springframework.security.web.savedrequest.SavedRequest;
 import org.springframework.stereotype.Component;
 import org.webcurator.common.ui.Constants;
 import org.webcurator.core.report.LogonDurationDAO;
@@ -32,7 +29,6 @@ import org.webcurator.core.util.CookieUtils;
 import org.webcurator.domain.UserRoleDAO;
 import org.webcurator.domain.model.auth.User;
 
-import javax.servlet.RequestDispatcher;
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
