+++ webcurator-core/src/main/java/org/webcurator/core/report/LogonDurationDAO.java
@@ -16,15 +16,28 @@
 package org.webcurator.core.report;
 
 import java.util.Date;
+import java.util.List;
+
+import org.apache.commons.logging.Log;
+import org.apache.commons.logging.LogFactory;
+import org.springframework.dao.DataAccessException;
+import org.springframework.orm.hibernate5.support.HibernateDaoSupport;
+import org.springframework.transaction.TransactionStatus;
+import org.springframework.transaction.support.TransactionCallback;
+import org.springframework.transaction.support.TransactionTemplate;
+import org.webcurator.domain.model.report.LogonDuration;
 
 /**
- * Interface for marking a User's login / logout, storing his/her associated session
+ * Class for marking a User's login / logout, storing his/her associated session
  * @author MDubos
  *
  */
-public interface LogonDurationDAO {
-
+public class LogonDurationDAO extends HibernateDaoSupport {
 	
+    private static Log log = LogFactory.getLog(LogonDurationDAO.class);
+    
+    private TransactionTemplate txTemplate = null;
+
     /**
 	 * User has logged in
 	 * @param sessionId User's session id
@@ -33,22 +46,123 @@ public interface LogonDurationDAO {
 	 * @param username Username
 	 * @param userRealName User real name
 	 */
-	public void setLoggedIn(String sessionId, Date loggedInTime, Long userId, String username, String userRealName);
+	public void setLoggedIn(String sessionId, Date loggedInTime, Long userId, 
+			String username, String userRealName) {
+		
+		log.debug("setLoggedIn: sID=" + sessionId + " in=" + loggedInTime + 
+				" uId=" + userId + " un=" + username + " uRN=" + userRealName);
+			
+		final LogonDuration ld = new LogonDuration();
+		ld.setSessionId(sessionId);
+		ld.setLogonTime(loggedInTime);
+		ld.setUserOid(userId);
+		ld.setUserName(username);
+		ld.setUserRealName(userRealName);
+		
+		txTemplate.execute(
+				new TransactionCallback(){
+					public Object doInTransaction(TransactionStatus ts){
+						try {
+							log.debug("Before saving of the LogonDuration");
+							getHibernateTemplate().saveOrUpdate(ld);
+							log.debug("After saving of the LogonDuration");
+						} catch (DataAccessException ex) {
+							log.warn("Setting Rollback Only",ex);
+                            ts.setRollbackOnly();
+						}
+						return null;
+					}
+				}
+		);
+
+	}
+	
 	
 	/**
-	 * User has logged in
+	 * User has logged out
 	 * @param sessionId User's session id
-	 * @param loggedOutTime Time of User's logout
+	 * @param loggedOutTime Time of logout
 	 */
-	public void setLoggedOut(String sessionId, Date loggedOutTime);
+	public void setLoggedOut(String sessionId, Date loggedOutTime){
+		
+		// Find associated login record
+		List results = getHibernateTemplate().execute(session ->
+				session.getNamedQuery(LogonDuration.QRY_LOGON_DURATION_BY_SESSION)
+					.setParameter(1, sessionId)
+					.list());
+
+		log.debug("setLoggedOut sId=" + sessionId + " found " + (results == null ? "null" : results.size()) );
+		
+		if(results.size() > 0){
+		
+			// Set logout and duration (in seconds)
+			final LogonDuration ld = (LogonDuration) results.get(0);
+			ld.setLogoutTime(loggedOutTime);			
+			ld.computeAndSetDuration(false);
+			
+			txTemplate.execute(
+					new TransactionCallback(){
+						public Object doInTransaction(TransactionStatus ts){
+							try {
+								log.debug("Before saving of the LogonDuration");
+								getHibernateTemplate().saveOrUpdate(ld);
+								log.debug("After saving of the LogonDuration");
+							} catch (DataAccessException ex) {
+								log.warn("Setting Rollback Only",ex);
+	                            ts.setRollbackOnly();
+							}
+							return null;
+						}
+					}
+			);
+			
+		} else{
+			log.warn("No login associated found for session: " + sessionId);
+		}
+		
+	}
 	
 	
 	/**
 	 * Make sure that all durations of previous logins are 
 	 * set for the current user. I.e previous session are 
-	 * all closed even if Tomcat has been shut down e.g. 
-	 * @param currentUserOid ID of the current User
-	 * @param currentUserSessionId ID if the current User's session
+	 * all closed even if Tomcat has been shut down e.g.
+	 *
 	 */
-	public void setProperLoggedoutForCurrentUser(Long currentUserOid, String currentUserSessionId);
+	@SuppressWarnings("unchecked")
+	public void setProperLoggedoutForCurrentUser(Long currentUserOid, String currentUserSessionId){
+		List<LogonDuration> results = getHibernateTemplate().execute(session ->
+				session.getNamedQuery(LogonDuration.QRY_UNPROPER_LOGGED_OUT_SESSIONS_FOR_CURRENT_USER)
+				.setParameter(1, currentUserOid.longValue())
+				.setParameter(2, currentUserSessionId)
+				.list());
+
+		for(LogonDuration ld : results){
+			log.warn("closing logonDuration: " + ld.getOid().longValue() + " (session: " + ld.getSessionId() + ")..." );
+						
+			final LogonDuration finalLd = ld.clone();
+			finalLd.computeAndSetDuration(true);
+			
+			txTemplate.execute(
+					new TransactionCallback(){
+						public Object doInTransaction(TransactionStatus ts){
+							try {
+								getHibernateTemplate().saveOrUpdate(finalLd);
+							} catch (DataAccessException ex) {
+								log.warn("Setting Rollback Only",ex);
+	                            ts.setRollbackOnly();
+							}
+							return null;
+						}
+					}
+			);
+			
+		}
+		
+	}
+	
+    public void setTxTemplate(TransactionTemplate txTemplate) {
+        this.txTemplate = txTemplate;
+    }
+
 }
